refactor: unify the code base, add validation
Since the code base baseicall does the same for four different objectives,
we shouldn't have multiple, similar implementations of basically the same things.

This introduces a generic config class that is used to register
permission, visibility, and validation classes. The interface to using the package has been
cleaned up as well: The model classes won't need a mixin anymore for either visibility
or permissions to work. The permission classes don't need to inherit from the `BasePermission` anymore either

There is now a clean import / usage structure to use this package:

* all permissions stuff can be imported from `generic_permissions.permissions`
* all visibility stuff can be imported from `generic_permissions.visibilities`
* all validation stuff can be imported from `generic_permissions.validation`